* index_queue can now take multiple field name values  * Oracle doesn't support multiple VALUES inserts  * Make more use of multi-value insert calls to insert()  * Added begin/rollback/commit methods (test in import_rand_data) - this helps Oracle performance
git-svn-id: https://svn.eprints.org/eprints/trunk/system@3388 9491667e-5006-0410-a446-efbe8990b998

+++ perl_lib/EPrints/Index.pm
@@ -307,7 +307,10 @@ sub _do_ordervalues
 				$ov;
 		}
 
-		$session->get_database->delete_from( $ovt, [$keyname], [$keyvalue] );
+		if( !$insert )
+		{
+			$session->get_database->delete_from( $ovt, [$keyname], [$keyvalue] );
+		}
 		$session->get_database->insert( $ovt, \@fnames, \@fvals );
 	}
 }
